Fixed incredibly annoying test bleed issue
Looks like the H2 database was not released by one set of tests but OrcaSmokeSpec then tried to re-create the tables. Had to add @DirtiesContext annotation to tests to ensure the application context gets destroyed at the end. Because I wouldn't want my tests to be idempotent or anything, would I? Fucking Spring.