Refactored the TestStage to use Tasks instead of Tasklets
This decouples some of the test implementation from the Spring Batch API and enables me to debug some weirdness around aborting stages when a task fails. It means more tests are going via the TaskTaskletAdapter as well which makes sense as *everything* will go via that in a real pipeline.