Refactored stage configuration to always use flow builders
The batch API here is pretty hairy and confusing. It's really frustrating that terminating the job after a step fails prevents the afterStep method of a StepExecutionListener from running. That seems like a dumb oversight. Especially since part of the point of that method is that it can change the exit status of a step – e.g. it could be used for recovery logic.