#82 stages now have outputs
Data returned by Task.execute is added to the stage's outputs. For now I've removed the different treatment of data depending on task status. I'm not sure we really need that & I'd rather add it in response to demand than provide unnecessary complexity.

The Pipeline & Stage object model needs tidying up so that a) it's all immutable & tasks can't do inappropriate things with context and outputs. b) it's not painful to create simple stages without pipelines in unit tests.