Fix for #712: Deploy Stage of Pipeline would reset to default state.
This issue was caused by the multiple watchers on command.capacity.desired causing other watchers to fire in the deployStage.js
that are meant to synchronize the state between the pipeline stage object and the command object.  This watcher-chain-of-death was
causing the entire stage object to be reset whenever the capacity was changed.

To fix this I replaced the watchers on the directive controller with an function call that just updates the state on an ng-change.