Fix for SPIN-316:  Adding the entire orchestration model to the OrchestrationViewModel.
Doing a red/black or highlander strategy when cloning an ASG results in the 'deploy.server.groups' variable being overwritten (showing the name of the disabled asg) when creating the OrchestrationViewModal. We need/want the name of the new asg name from the clone command.
The solution will be to attach the complete Orchestration model to the OrchestrationViewModel. This will give us access to the entire task model where we can pull off the correct 'deploy.server.groups' value. In talking with Tomas this will increase the size of the data going over the wire but will allow access to the entire Orchestration model.