1st pass at ripping out batch persistnce
we now use a MapJobRegistry but we may not even need that. I still need to remove the trickery around resetting job execution status before restarting but it might be that we don't even need it as we can just start a new execution that is smart enough to skip completed steps. There is some nastiness in ExecutionStarterSpec that exposes some tangled dependencies but other than that this feels a bit lighter.