found typo which suggests we have a missing test so I wrote the test
The previous version was trying to do a full delete on an entity which it already knew didn't exist so it was inevitably going to just throw an exception. I don't think it's a big deal that we don't run a full cleanup in this (very unlikely) scenario so I've stuck to just removing the erroneous set entry.