Fix for a problem where > 1 synthetic stage of the same type with the same owner will be treated as duplicates
This code was supposed to be dealing with re-building pipelines for restarts but it's breaking when > 1 synthetic stage of the same type needs to get inserted. The id generation is kinda dumb here and I'm going to try to fix this a better way but it's urgent we get something in as this is probably breaking several pipelines.