Temporarily added spinnaker tests from private repository
These are not intended to live here long term, but putting them here makes
it easier to share with Netflix and easier for them to consume. I'm intentially
not putting it in the citest package because that may make it harder to
separate them back out later. This way they are still separate from python's
perspective, but are the same from github's.