Refactored and migrated scripts from private google repository.
These are deposited under a "google" directory. The intention is that
they will be refactored over time to generalize and moved upward in
the filesystem. The "google" directory could still provide a means for
google-specific overrides or extensions if they are needed and not
appropriate for later. At least for the time being, this seems like a
non-intrusive way to make these scripts available now, without worrying
about generalization, and let us continue to use them as they are generalized.

See google/README.txt for more info.

NOTE: configuration parameters names changed, renaming GCE_ to GOOGLE_
and adding a GOOGLE_ prefix to google provider-specific parameters that
were missing it.