Allow Cassandra and Redis to be configured to run other than localhost.
Cassandra and redis are currently still installed but they will not be
started up if the configuration says remote. The installation of these
is not affected, so they are still run locally when run. On the positive
side, this means you can currently easily change the configuration back to
localhost. But this isnt useful in the long run.

There is no installation support for a remote cassandra and redis.
It is assumed to already exist or be provided externally
Presumably you could install these from launcher, though I did not get
that to work on my first attempt.

In order to get this to work/verify, I installed
a second VM with spinnaker, then configured redis and cassandra to bind to
their internal IP address and started just them up (i.e. it was a vanilla
redis and cassandra installation). Then I conigured the remote
spinnaker_config.cfg to point to it as the CASSANDRA_HOST and REDIS_HOST
and ran start_spinnaker.sh as normal. There is no redis or cassandra running
locally and spinnaker starts up (and configures the remote cassandra schemas).

ElasticSearch is not included here. A followup CL is going to remove it.