Write placeholder into new directories to force them to exist
It turns out that the existing script does not preserve the directory
structure if the target directory does not already exist, and a straight up
copy from one dir to another would produce an extra level of directory,
which isnt desired. This solution forces the directory to exist so that
the copy retains the original structure.