Handle process-level test errors.
Tests currently return 0 rather than non-0 on error.
This changes the test return result to be non-0 on error
(value is #failures + #errors but that isnt officially defined yet)

Also, force cleanupScenario to be called on exceptions too.

Made two unrelated changes while I was here:
Updated spinnaker images to the most current ubuntu (not a long term solution).
Gave default value to spinnaker initArgumentParser system name so that this
  isnt actually required.