Updated config to specify host/port for spinnaker services.
This is so they can be bound to localhost for enhanced security.
And so it is easy to rebind the ports.

Dont set environment variables if they were already set,
and use the environment variable as the actual value for override.

Variables can be set off other variables transitively.

Also added run/fetch modules to start factoring these out of the installer
to be able to generalize the scripts. For now the old implementation is
still there and untouched because lots of other modules use it. This is
the first of several files waning off it.

Updated the run scripts to consider the host binding when checking up.
Also the script to the updated run/fetch apis.