Refactored base test class into test_runner
Removed scenario from base class and introduce it in service_testing for the time being.
I'd like to refactor that further in a different CL.

The runner is also responsible for reporting. I expect to refactor that out in the
future as well, but this is a first step to make test classes lighter weight and
remove the shared global state at the foundation.