Allow CachingAgents to explicitly provide evictions.
Our agents that cache on_demand items need to clean those up when they expire. This doesn't line up with
the authoritative / informative notion for agent data sets.  Instead of the caching agent manually calling
evictDeletedItems while building a result, this moves the eviction set into the CacheResult so that it
can be deleted during (but after) the cache write for a data type.