Encapsulate the "Spinnaker" release package as a tar file.
This does not include the subsystem debian packages, only the spinnaker.git
project artifacts.

This is accomplished by having build_release generate a tar file and copying
that to the --release_path rather than building a filesystem. The install
script in turn copies and untars this rather than individual files.

The intent is that the debian packages would get written to bintray
(or some apt-repository source) not the --release_path. That leaves just
the tar file, (and new install_spinnaker.sh) which should eventually be
packaged together as a debian package or could be placed somewhere
(e.g. bintray) for an easy curl download. The placement and easy download
will be addressed later.

The interfaces for building a release and running the installer are
still unchanged from before, however this change does not provide backward
compatability for the installers. That should not matter since the
--release_path and installer area already paired in practice.