Add user menu to netflix mode enabled
Add the user menu when netflix mode is enabled. The user menu will show up
on the right side with the whatsnew and feedback icons. I moved the icons into the
actual navbar with the search menu so it worked like the user menu. I overrode the navbar-right slightly
for the header so it could support more items (known issue with bootstrap http://getbootstrap.com/components/#navbar-component-alignment)