Adds a migration path from legacy UDF to newer UDF for AWS UserData generation
The legacy UDF mechanism supports assembling a UserData script customized by account/application/
environment. We are migrating away from this (it is generally not a good idea in the spirit of
immutable infrastructure to have logic in UserData), to a single UserData template file in which
we just token replace environment variables. The logic to customize behaviour based on those
values is baked into the image.

An application can explicitly opt in or out of the legacyUdf format by setting a legacyUdf flag
in the application metadata in front50. Additionally there is a global default when the preference
is unspecified for the application is configured via udf.defaultLegacyUdf property