Avoid evicting cached server groups that are newer than what Edda knows about
Backstory:
It is possible for Edda to return a set of server groups with an `mtime` of N that do not accurately reflect the true state of AWS at time N.

Workaround:
Whenever we see an Edda result that appears to be newer than a previously ForceCache'd result, fetch the server group directly from AWS and compare a few attributes:
- desired/min/max sizes
- suspended processes

If the direct read and edda match, we are free to evict the ForceCache'd result.

If they do not match, we'll keep the ForceCache'd result.