SPIN-1415:  Evict Local Storage if there is an exception on setItem
Our current strategy to only clear out the infrastructureCaches if there is an exception thrown on insert.
As the app has grown we have added more and more to Local Storage resulting in the possibility of exhausting the 5MB of space given in most browsers.

This change simply evicts the entire local storage (restricted to the namespace of the url), when there is an insert exception.

This change also reduces the TTL of the 'collapsible section cache' from 30 days to 7 days.