Renamed verifier contains* excludes* methods
The names arent as concise, but might be easier to remember and distinguish.
  contains -> contains_path_value
  contains_eq -> contains_path_eq
  contains_pred -> contains_path_pred
  contains_group -> contains_pred_list
  excludes -> excludes_path_value
  excludes_eq -> excludes_path_eq
  excludes_pred -> excludes_path_pred
  excludes_group -> excludes_pred_list

These are still not ideal. "contains_path_value" and "contains_path_eq" sound
similar. contains_path_eq is looking for exact equality, and isnt that common
for us. The contains_path_value is really "contains a path whose value contains"
and is far more common because we dont know literal string choices and often
the value is composite and the test only cares about a subset of attributes.
I could call this "contains_path_subset" but "subset" isnt really the right
term for strings, which is very common.