Adds support for starting a pipeline from a webhook POST.
Webhook support expects a JSON payload in the POST.
This JSON is a map of items. These items are compared
to a map of "constraints" in the Trigger configuration.

If all the constraints have matching items in the payload
map, the pipeline trigger fires.

Constraints that have a name and value must have an exact
match in the webhook payload.
Constraints that have a name but the value is an empty string
are considered met as long as there is an item in the payload
with the same name, regardless of the value of the item in the
payload JSON.