makes ContextParameterProcessor explicit about behavior of unknown attributes.
Adds an allowUnknownKeys option when processing parameters - when transforming a pipeline before execution
this is set to false to ensure any unresolved expressions do not get interpreted. This addresses the case
of an expression as part of a larger string such as `${foo}-bar` getting resolved to `-bar`.

Transformation inside the execution of a stage allows for missing keys which enables the use of the SpEL
elvis operator and preserves existing behavior.