Parse a FQPN at any position in the package list (#88)
Parse a FQPN at any position in the package list.
This patch is a step forward in correct handling of multiple packages. With this patch, if a FullQualifiedPackageName gets passed in the
package list it gets parsed no matter the position (before it only happened for the first one).
Also, the first package keeps its importance of potentially determining the ami-name or the Appversion if correctly parsed.