Add publicBakeryBaseUrl
View bakery log link in Deck always point to bakeryBaseUrl. So, if for deployment reason the bakery has two ursl. One public and one used internally, behing a LB, the links will point to the internal one. So, this links will not work when trying to access them in Deck. This change allows you to specify a public bakery url to be used by deck. If it is no provide, it will fallback to baseBakeryUrl

Recover some missed vars

// var gateUrl = ${services.deck.gateUrl};
// var bakeryBaseUrl = ${services.deck.bakeryUrl};