Fixes SPIN-1797
Initial load of the vpcTag was showing "EC2 Classic" as it's label if the vpc data was not pulled down.  Once it was loaded the correct subnet would show.  This was leading to some confusion to users who had recently upgraded to VPC0.

Removed the default "EC2 (Classic)" label and replaced with a spinner.