SPIN-1900: Track scrollbar position on pipeline graph (#2461)
* SPIN-1900:

Cadmium has a very long pipeline graph that and the overflow-x: auto kicks in.

Due to the change in the fetch interval to 1 second, when every the user scrolls the to the right when the graph repaints it resets the letfScroll on the graph to zero, creating a very frustrating user experience.

This seem to only happen when the pipeline is Running.

This fix adds a mousewheel event (scroll was not firing) on the <pipeline-graph> element.  It then saves the scrollLeft posisiton of the scrollbar off to the pipelineGraphService.

When the svg graph redraws it get the scrollLeft offset from the service and sets it.  This prevents the scroll bar from jumping back to the zero position every second.

* remove console.log