Fix for Canaries Timing out
There is an edge case when a canary is with both a duration and warm-up periods are set the canary times out before completion.
Basically there was no padding built into the stage timeout time.

This add a 2 hour padding to the time out to allow for the canary to complete and report back before timeout.