Allow InstallSpinnaker.sh to ommit cassandra (#1026)
When omitting cassandra, it will still add the package manager location
but will not actually install it. I can remove that too, but figured I'd
wait until we completely remove cassandra. In the meantime adding the location
will make it easier to manually install cassandra later should the deployer change
their mind.

The default behavior for GCP is to leave out cassandra. The other platforms are unchanged
for the time being. Adding a --noinstall_cassandra will force it off and --install_cassandra
will force it on.

Removed explicit cassandra (and redis) management entirely from developer mode (running from source).
To use cassandra from source, you'll need to ensure it is running yourself.
If it is an upstart service (as is normal) then it will already be running anyway.

I left the stop scripts in for now because I'm not sure if anyone is using the stop dependencies
script, but expect to remove this as well.