Added base.ExecutionContext
This is a basic API for now, but will likely change over the next CL or two.
An execution context is essentially a dictionary, but can have some of its
elements snapshotted. The dictionary is passed through APIs in the test
execution flow, acting as a data store for parameters so that runtime data
values can be exchanged among different methods using predetermined keys,
thus implementing variables. This can be used to specify contracts involving
specific resources generated in the operation whose exact names are not known
ahead of time.

Initially the context is only used in contract validation, but that is just
to keep the CL smaller. It is an optional parameter in the run_test_case
method used by every test so that we dont need to modify every test in this
CL. However eventually these tests will explicitly pass in the context and
the parameter will become required.