Adding default configs for OpenStack.
The default Openstack configuration will now read from the same environment
variables that are standard in the Openstack RC files. Spinnaker doesn't
support reading these files directly due to these files being
interactive by default to prompt a user for a password. But using the
same variable names gives a smoother way to utilize this file to setup
the environment for Spinnaker.