Add new deleteBakeByKeyPreserveDetails() BakeStore entry-point and em… (#140)
Add new deleteBakeByKeyPreserveDetails() BakeStore entry-point and employ it when rebaking. This allows older bakes that have been rebaked to retain their logs and other details.
Modify deleteBakeByKey() to return the bake id of a deleted bake and include this in the response message.
Modify RedisBackedBakeStore.deleteBakeByKey() to resolve the appropriate rosco instance id instead of assuming the bake is being rebaked on the same rosco instance.
Modify BakeryController.createBake() to also cancel the executor's bake job when performing a rebake.
Add script comments and drop string literals in favor of enum.