Can't nest ${..${}} in config
This won't fail to parse, but it won't resolve the nested ${} expression. I'm dropping the env reference, since it's more likely that `primaryCredentials` is correctly defined. We can talk about a better way to resolve which project handles stack driver, but I figure we should leave this in an intermediate, working state.

There is another fix to kork incoming separate from this that is actually causing everything on kork 1.77 to fail that has stackdriver enabled (so far clouddriver, echo & front50).

@ewiseblatt, @duftler PTAL