Rewrote stackdriver integration
Instead of being optimistic that I can create informed Metric Descriptors specifying
the labels that will be needed for the different use cases, this approach is
pessimistic and assumes that the definition is eventually going to be discovered
to be wrong and fixes the definition when it encounters a problem.

I'm not convinced that this will not lose data -- it does not at the moment,
however the issue exists either way. At least this strategy adds resilience
and makes it easier to have a judas deployment easily flush out any new labels.
The underlying descriptors can be extracted to proactively define them for new
deployments as a better starting point.

This removes the need for actually caching descriptors or having "hints" as a
workaround.