Restrict cross origin requests.
*Note:* this is a breaking change for anyone that is currently relying on the
cors filter to allow any requesting domain with authentication, and is hitting
the spinnaker API directly from other browser based web apps besides deck.

Adds a new `cors.allowedOriginsPattern` configuration property that takes a
regex to match against the request origin. If successful, the request origin is
returned as the `Access-Control-Allow-Origin` header. If unsuccessful, `*` is
returned and the browser will then not supply cookies.

If not configured, then falls back to checking the hosts configured in the
`services.deck.redirectHostPattern` property. Finally falls back to matching
`services.deck.baseUrl`.

A reasonable example value for `cors.allowedOriginsPattern` would be:

````yaml
cors:
  allowedOriginsPattern: '^https?://(?:localhost|[^/]+\.example\.com)(?::[1-9]\d*)?/?$'
````

This would match any `http` or `https` request from `localhost` or `*.example.com` with or without a port supplied.