Clean up default docker config
The default clouddriver.yml ignores the repository specified in spinnaker.yml and instead gets the entire catalog. This is confusing because some registries (like the default 'index.docker.io') don't let you get the entire catalog and return a '401 Unauthorized'. I modified clouddriver.yml to use the repository, but also removed 'library/nginx' so that it still gets the entire catalog by default

Add ‘:’ to the username and passwordFile environment vars so that they default to null (and thus you can connect to public repos). Otherwise it was using the literal string “${SPINNAKER_DOCKER_USERNAME}” as my username