(core): pull angular-cron-gen library internally
* pull angular-cron-gen library internally to deck to fix several issues
with the library.
* the latest version of the library will not allow angular expressions
for the new `template-url` property as it uses the `templateUrl`
function which executes before the scope is initialized so any
scoped property value is not accessible.  this means that our template
value set via a controller property is not accessible.  this is a
long-standing issue [in the core library](https://github.com/angular/angular.js/issues/2895)
and it doesn't look to be fixed anytime soon.
* an hourly cron trigger set in the UI to execute start at a specific
time actually generates a daily cron expression so the portion of the
hourly cron trigger that allows starting at a specific time was
removed from deck's cron picker template.
* the regex used by the library to match hourly cron expressions was
incorrectly matching daily cron expressions which was resulting the UI
displaying `hourly` instead of `daily` in the select so the regex for
hourly was fixed to properly test for matching hourly cron expressions.