Make the halconfig thread local (#139)
There was a corner case preventing validators from using the lookup
services, since they would implicitly load from disk rather than
validate what had been modified in memory. This was caught in the
edit-storage command which would rely on the account lookup service to
check that the specified account was of the correct type.