(netflix) Allow Fast property scope to be updated. (#3375)
Once a Fast Property has been set the scope can't be easily updated b/c the key of the
Property is a pipe demilted string of the property key and all the values of its scope.

To get around that limitation we create 2 Persisted property stages.  The first one will
create a new property with the new scope.  Once that completes we follow up with a delete property
stage that cleans up the origional property.

When editing the property the user will be shown the original scope and they will be able to select the
new scope.  If they choose to to update the scope they can remove it and only an upsert property stage
will be used.