fix(dev): clean google tarball more
This fixes a bug with the clean_google_image script and
also applies it to the google image tarballs.

To make it used for both, I refactored the build_google_image script so it
can preserve the disk and/or emit tarballs instead of or along with images.
This obsoletes the google/dev/build_google_tarball script, but am keeping
it around until other processes that use the script switch over.