fix(tests): Fixed native test startup / configuration
Get configuration parameters from the "resolveEnv" endpoint.
Show exception information if agent fails
If not given a location, assume localhost

The net effect is that if you are on the same machine as spinnaker,
and the tests primary microservice is on its standard port, then
you can run a typical test without any command-line parameters provided
it can get whatever configuration it needs from the server.