Remove AWS Variables
Removing the AWS environment variables. Ran into an issue where the
variables were also defined in the ``compose.env`` file and were being
overridden by the `docker-compose.yml` and ended up being empty. This is
confusing for new folks trying to use Spinnaker with Compose as it seems
like the variables aren't being set if they're defined in the env file.