fix(validate): fixed location assumptions in tests. (#1741)
This still doesnt work in general for aws, though
it does work with my paritcular setup. My intent
here was to be able to easily change the region,
but I still cannot do that. I think this is still
an improvement overall, though doesnt go as far as
I wanted.