feat(provider/kubernetes): Run Job extended config (#3823)
Adds the ability to add labels and annotations to the Run Job pod
definition. This is useful for pin pointing pods for logging
purposes. I've transitioned everything from using the
`kubernetes-annotation-configurer` to the core `map-editor` for
consistency. I've also modified it to provide options for displaying it
in a manner similar to what `kubernetes-annotation-configurer` looked
like.