feat(core): AFTER_STAGE stages can now be planned post task completion (#1843)
This allows for interesting scenarios where a stage could have a `Task`
responsible for building the context that is subsequently used to build
a dynamic set of synthetic stages.

Currently this logic must reside entirely in `aroundStages()` and you
lose the ability to arbitrarily retry with backoff's, etc. that normal
`Task` execution provides.