This adds the ability to restrict github authentication to only users (#490)
belonging to a certain organization. It adds in a provider specific bean
to encapsulate the parts that aren't general oauth2 handling. The
interface is defined in SpinnakerProviderTokenServices, and it's just a
single call that takes the user token and the map of user info. The
github provider bean keys off the
security.oauth.providerRequirements.type setting being 'github', and if
so it loads the extra provider specific requirements.  This is a
proposed fix for spinnaker/fiat#188