feat(buildtool): Added publish commands for changelog and apidocs
This does not publish the changelog as a gist as before, though does
force it through a PR now.

This also refactors the changelog generation to get rid of the details
(by default) and to group changes by component so it is more concise
and easier to read.

The apidocs are written as before, not with the versioning changes I
proposed earlier. I'll make those changes later as well.