feat(webhooks): support preconfig params (#1887)
adds support for configuring predefined parameters as part of a
preconfigured webhook stage. this will allows admins to add options for
users to fill out. the values set by users can then be used in the other
options via SPEL. this is a really basic implementation and only
supports string types options.