feat(kubernetes): Allow custom pod annotations for spinnaker services (#861)
When deploying spinnaker to kubernetes, it is sometimes necessary for
some of the services to have extra annotations. For example when using
kube2iam to specify the role certain pods should get. This change adds a
map for podAnnotations to the KubernetesSettings and applies it to both
the replicasets in the v1 provider and the deployments for the v2
provider.