feat(aws): V2 security groups (#63)
Removed the generic security group intent processor from `keel-intent`.
Moved nearly all of the logic into `keel-intent-aws`. Should other cloud
providers come in, we'll look at how to make the functionality more
generic.

The v2 implementation differs from v1 in that security group intents can
define an entire security group, or just one or more rules on a security
group, as well as rules on a security group that is not managed by keel.

To illustrate why this is important: If we have a security group (SG-1)
owned by Team A and Team B wants their security group (SG-2) to have
access to it, they will need to modify SG-1, which is owned by a different
team. Requiring a team to manage inbound change requests to grant access
to their services is an undesirable coordination point, and one that has
been avoided in Spinnaker to date.

This also serves as a good base for Authorizer, which will allow people
to describe their service dependencies at an abstraction level higher
than security groups, and we'll be able to manage both ingress and egress
rules for users.

This work exposed a few things I still need to add, or refactor in Keel:

1. Need a way to allow teams to create a policy that restricts who can
   actually use rules against a security group. Will add once there's an
   ask, but it is an inevitability.
2. I need to remove the user-defined intent IDs. That feature will add
   far too much complexity early on for edge cases that we don't need to
   deal with yet.
3. I need to figure out a better way to create internal attributes &
   labels for intents. I don't like that I create a user-space label for
   managing parent/child associations.
4. We're going to be converging state for every rule intent every cycle.
   There will need to be a better mechanism in the scheduler for deciding
   what intents actually need to be converged and ones that will be
   implicitly applied through association.