feat(buildtool): Write changelog gists
This also changes publishing spinnaker to use changelog gists.
There is a process change here where the publish directly writes
the changelog entry to spinnaker.github.io, but takes the existing
gist URI as a parameter as discussed offline. This gist is assumed
to already be curated so no need for additional review step.

The push_changelog_to_gist introduced here is a new process step
where the changelog is uploaded to a separated volatile gist
(control outside this script) which is used offline while curating
the official changelog gist.

Conflicts:
	dev/buildtool/base_metrics.py