fix(travis): Fix a regression that cause getBuilds() to be slow (#239)
The code was always _supposed_ to fetch logs in this code path. However, due to a bug in the JSON to Groovy deserialization, the field containing job id's was always null. Due to groovy's nature, this just caused Igor to not fetch any logs, instead of failing.
I fixed the issue, but then realized that the log fetching made the code way slower in places where we didn't really need artifacts. So I've disabled logs/artifacts fetching from everywhere but the TravisBuildMonitor.