feat(tagging): Select which upstream stages to include in image search
When tagging an image, the default behaviour is to look for images in all ancestor stages. In some cases, this is not the desired behaviour. This commit makes it possible to optionally define a list of stages that will be considered in the tagging stage.
This commit corresponds to the following orca PR: https://github.com/spinnaker/orca/pull/2138