fix(citest): Add queueing to gate Hystrix config
A number of the intermittent failures observed in running our CI tests
are due to Gate returing a 500 on /tasks/task_id because Hystrix did not
have an available thread to use to query Orca.  As this is primarily a
problem with requests being bursty rather than an overall throughput
problem, adding a modest queue to the Hystrix config so that Gate can
handle and recover from short bursts of queries.