fix(core): Remove bashisms from start.sh
It's explicitly started using `sh` in `package.json`, so it should
not use bash (although it has `#!/bin/bash`).

The alternative is to change `package.json` to use `bash` instead,
but since this script has no need for an advanced shell, just use
POSIX shell features.

This makes `yarn run start` work on systems where `sh` is not bash,
where it previously failed with the error message

> ./start.sh: 8: ./start.sh: [[: not found