fix(citest): Fix race condition in triggering pipeline
In the bake and deploy test, we create a pipeline, wait for
front50 to confirm it exists, then try to trigger the pipeline.
This sometimes fails, because it can take up to 10 seconds for
Echo to cache the new pipeline, and trying to trigger it during
this time will be a no-op.