fix(citest): Increase timeout for resize server group
When resizing a server group, Spinnaker reports success when
the new instance is observed. Our test cases are actually testing
that the instance group is reporting the new size, which can lag
the instance actually appearing. Increase the amount of time we
wait for the instance group to report its new size to reduce
intermittent failures.