Add the commands needed to run the kayenta tutorial
The tutorial is currently providing two ways of creating
the "Automated Canary Deploy": with the GUI (the main way,
where the reader learns the most), and with the CLI (backup,
in the case the reader messes up their config).

The editors are asking to move the backup path outside of
the main tutorial document. So I'm putting it here, in the
README, with the commands needed to create the 2 other pipelines.