fix(moniker): handle Monikers without cluster (#2433)
Cluster is not a required field on Moniker, and in a number of cases where we are working with a Moniker from
pipeline configuration or a submitted execution the cluster can be absent. For TrafficGuard, locking, and
cluster status checking purposes we want to programmatically rely on the Moniker object holding the cluster
to check. This adds a few more checks and fallbacks to the serverGroupName to build the Moniker object if
there is no cluster available