fix(core): Ensure event processing occurs in the same thread
Currently we guard against processing events before the pipeline
cache is initialized (and thus having the events be no-ops) by
using RxJava's zipWith to wait until we have both an event and
a cached list of pipelines.

This means that the actual downstream processing will be handled
by whichever thread finishes last; in most cases this will be the
thread that called processEvent, but in cases where the pipeline
cache is not initialized, the event will actually be processed
by the pipeline refreshing thread.

Now that we're propagating authentication headers through Echo,
we need the event to be processed by the thread that has those
headers and can't have it processed by the pipeline refresh thread.
Let's instead create a getPipelinesSync method that blocks if the
cache is not initialized, and have the event processing thread
block on that function.