feat(ec2): Change cardinality of port ranges on security group rules
Previously a rule could have multiple port ranges but this didn't really reflect what was happening under the hood on EC2 and it made it hard to diff current and desired models. For example, a security group would appear to have a diff because River shows it as having a single rule with 2 port ranges where the desired state is two rules with a port range each. Now we normalize the data coming out of River and require the desired model to have a single port range per rule so there's no ambiguity.