fix(telemetry): Unwrap InvocationTargetException (#205)
The change to use an instrumeted proxy for Jenkins calls is masking
error codes received from the Jenkins server. This is because exceptions
thrown in InvocationHandler.invoke are wrapped in the checked exception
InvocationTargetException before being returned to the caller. This
ultimately leads to downstream code receiving an
UndeclaredThrowableException rather than the particular exception.