feat(scattergather): Initial scatter/gather library (#3173)
Development implementation of a scatter/gather library, for future use in
clouddriver federation, allowing operators to shard write & read traffic
without requiring development work on other services. It's possible this
behavior could be lifted to its own federation service in the future if
other services need the same behavior.

This API is currently coupled to `HttpServletRequest`, which it will then
fanout into a collection of OkHttp `Call` objects that will be executed
according to the provided `ScatterGather` implementation. Currently only a
`NaiveScatterGather` is available, which performs the calls sequentially
and without any error handling: This is for development purposes only.

Followup PRs will be made to add a `ScatterGather` implementation that
executes requests in parallel, performs error handling & retries as well
as cancelations in the event of timeouts.

Responses from shards are reduced into a single `ReducedResponse` by a
pluggable `ResponseReducer`. The existing implementation is a simple
deep merge reducer.

Better to place this in kork? There's nothing specific to clouddriver in
this library, nor do I expect that there ever will be.