refactor(controller): Refactor common logic in controller GET handling (#1116)
* style(webhooks): Fix indentation in WebhookController.java

* refactor(controllers): Add GenericGetRequest module

Adds a module GenericGetRequest that handles the common code that
most controllers duplicate.  Also adds ValidationSettings as a
model attribute so the code to handle these query paramters can be
factored out of the controllers.

* refactor(controllers): DefaultControllerValues -> DefaultValidationSettings

* refactor(controller): Refactor common logic in controller GET handling

Replace the duplicated code in the controllers with calls to the new
GenericGetRequest class.  This should make it conceptually easier
to add new controllers for halyard settings, as it's now only
necessary to specify the getter, validator, and description of the
task, and the common logic is handled by GenericGetRequest.

* fix(controllers): Fix minor issues