feat(tests): add some functional tests to Deck (#6198)
This is the first in a series of commits adding functional testing
in Deck.

This commit adds two simple tests: a smoke test that opens the
home page and a test that checks for expected values in a GCE
clone modal.

These tests will only reproducibly pass if the developer sets up a
Spinnaker deployment and GCE infrastructure matching those used
when the test was written. A follow-up PR will add fixtures to remove
this limitation.