fix(docker): Update cache when a null tag becomes non-null (#344)
* test(docker): Add unit tests for shouldUpdateCache

Also lightly clean up some of the existing tests

* refactor(docker): Change return type and name of shouldUpdateCache

To prepare to allow cache updates without sending events, change
the return type of shouldUpdateCache to an object, and also
change the name to getUpdateType.

* fix(docker): Update cache when a null tag becomes non-null

We currently don't update the docker image cache when either
the prior or the current digest is null. This means that any
tag that was present before trackDigests was enabled will
always have its digest set to null. Change this so we do set
the digest any time the new one is not null.

In order to avoid triggering multiple pipelines when turning
trackDigests on, we won't send an event to echo when updating
from a null to non-null digest (as we don't actually know if
it was updated), we'll only update the cache.