fix(notifications): Support multiple app-level notifications (#2622)
Du to a bug in the comparison code, similar notifications of the same type was stripped away, so that only the first notification would be notified. In our case it was because `publisherName` was `null`, which made the `publisherMatches` criteria `true`, even though it was supposed to be `false`. I've fixed this by only considering non-null values.