feat(artifacts): Human readable expected artifact display names (#6344)
Generates readable names like "friendly-deer-94" as soon as an expected artifact is created. This name is editable by the user. The purpose of generating the name is to avoid making display name a required field needing validation wherever expected artifacts are created.

Expected artifacts are still referred to by GUID everywhere in the pipeline so that changes to the display name do not alter these relationships. 

To retrofit existing pipelines, we generate human readable display names in a way that is designed to contain elements of the artifact with a higher information density.