chore(build): Use newer base image for deck build (#6538)
The 'java' docker repo has been deprecated and not maintained
since the end of 2016 in favor of the 'openjdk' repo.

The old ones were using openjdk as well, so that's not a change,
just the repo.  This also makes the base debian image more modern
(stretch instead of jesse).

The particular issue this fixes is that deck wasn't building on
debian stretch after node was upgraded.