fix(build): Pass --frozen-lockfile to travis builds (#6611)
By default, yarn will update yarn.lock if it's out of date with
package.json. We don't want this to happen during CI builds, as
the CI system is not committing yarn.lock and so is testing an
unreproducible state. Yarn recommends passing --frozen-lockfile
on CI builds for this reason. This will cause Travis builds to
fail if someone bumps dependencies in package.json, but forgets
to commit yarn.lock.