feat(core): Use gradle to generate docs (#1225)
The Halyard docs are currently generated via a makefile. Move this
to gradle so that it can be incorporated into the build process
rather than a separate shell command.

In particular, the check task should fail if the docs do not
generate properly or if the docs committed to git do not match
what are generated from source. This will cause our Travis build
to fail and block merging when a PR either forgets to update the docs
or breaks doc generation.