refactor(jenkins): Move logic from BuildController to JenkinsService (#378)
* test(ci): Remove implementation checks from tests

Many of the test on the CI classes test the implementation of
the classes rather than the interface---in particular by verifying
the number of calls to buildMasters. Replace these tests with an
actual BuildMasters populated with test data and remove checks on
how many times its methods are called.

* refactor(ci): Improve the contract around build services

The BuildMasters class currently has a public map, which is accessed
from everywhere. Give it an interface, addServices() to add services
and getService() to fetch a service, and made the implementation
details private.

* refactor(jenkins): Convert JenkinsService and BuildService to Java

* refactor(jenkins): Move logic from BuildController to JenkinsService

The BuildController has a lot of logic that should be in the individual
build services. Move towards making this better by pushing some of the
Jenkins-specific logic into JenkinsService