fix(artifacts): Extract artifacts before enforcing constraints (#478)
* refactor(triggers): Push artifact extraction/matching to base

All of the event handlers duplicate the same logic around extracting
some artifacts from an event, and checking any artifact constraints
on the trigger. Push this logic to BaseTriggerEventHandler.

This commit also removes any need for the specific EventHandler
classes to know about Pipelines at all; they deal only in Triggers
and Events.

* fix(artifacts): Extract artifacts before enforcing constraints

We currently extract artifacts using Jinja templates as a final
post-processing step on the pipeline. This means that the artifacts
are not available when we check artifact contstraints. Fix this
bug by moving artifact extraction to occur during trigger evaulation.