fix(testing): Fixed gcs pubsub operation
A few changes here.
First, rather than rebinding the operation midstream from the gcs pubsub agent
to the gate agent, construct the operation with both agents and use the pubsub
agent to handle the pubsub operations, but bind the gate agent to the operation.
This is because the gate agent is used to check the status, which is what the
bound agent is for.

The PubsubTriggerOperationStatus seemed to have too much state.
It mostly delegates to the __trigger_status instead. Its only real added
state is whether or not it timed out.

Clarify between the trigger status and the trigger response to avoid confusing
the two.