chore(jinja): Upgrade jinjava (#2717)
The ModuleTag class breaks with jinjava >= 2.2.8 because of a fix in
failsOnUnknownTokens.  We need to catch this error in cases where we
expect that a token may be unknown. The test also uses True as a literal
which should be true (ie, lowercase) which worked before but breaks with
the upgrade.