feat(artifacts): Re-use artifacts when re-running a pipeline (#6663)
* refactor(artifacts): Rename ArtifactList to ResolvedArtifactList

This component is specific to a list of resolved artifacts, and
requires the pipeline's expected artifacts as input. Rename it to
free up ArtifactList for a more generic list of artifacts.

* refactor(artifacts): Create a react component for an artifact list

The logic for rendering a single artifact is currently embedded in
ResolveArtifactList; pull this into its own component so it can be
reused and create a new ArtifactList component that generically
displays a list of artifacts, without expected artifact logic.

* feat(artifacts): Re-use artifacts when re-running a pipeline

The manual execution widget does not currently support injecting
artifacts into a pipeline. This means that pipelines that depend
on artifacts cannot be manually started; support this workflow
for the re-run pipeline case by re-sending the same artifacts as
the initial run. These will show up as a read-only list of artifacts
in the Confirm Execution dialog.