feature(ci): Fetch artifacts from CI builds (#2723)
* refactor(ci): Convert BuildService and IgorService to Java

Both of these files are trivial to convert to Java. Also, in
IgorService, replace deprecated @EncodedPath with the equivalent
@Path(encode = false).

* refactor(ci): Clean up inheritance of CI stages

It's confusing that Travis and Wercker stages extend Jenkins stages;
make an abstract CIStage that these all extend. Also convert these
stages to Java, and clean up some of the string interpolation that
was hard to read.

* feature(ci): Fetch artifacts from CI builds

Jenkins triggers support inflating artifacts from the build results
based on a Jinja template specified in the build's properties. Add
the same functionality to Jenkins stages.

Use the property file defined in the CI stage to extract artifacts
from the CI stage build.

* refactor(ci): Pull getProperties into its own task

The prior commit added general support for retrying tasks that
communicate with Igor; update getProperties to be its own task
and have it use that new support.