test(ci): Add tests to JenkinsStage and ScriptStage (#2768)
* test(ci): Add tests to JenkinsStage and ScriptStage

There are no tests that waitForCompletion is properly read when
starting a Jenkins stage; this functionality recently broke, so
add some tests to it.  Also add some tests to verify that binding
of artifacts works correctly.

The ScriptStage has no tests at all; add a simple test that it
fetches the properties after the script finishes.

* refactor(ci): Move CIJobRequest and rename it CIStageDefinition

This definition was originally only used one specific task, but
it really represents the full stage definition of a CIStage, so
rename it and move it to a more appropriate location.

* refactor(ci): Remove explicit casting from CIStage

In order to quickly fix a bug, I explicitly added some code to
coerce waitForCompletion to a boolean. Let Jackson handle this by
adding it to the CIStageDefinition model, and also do the same with
the pre-existing expectedArtifacts field.