fix(triggers): Add timeout to polling on manual trigger (#6707)
When polling for the appearance of a new pipeline, we poll
indefinitely.  We should add a timeout; the new pipeline will
still appear due to the scheduled refreshes even if we time out
this poll.

This commit also changes startPipeline to chain promises so that
errors anywhere in the chain are caught and cause us to unset
triggeringExecution (which causes the component that triggered
the execution to return to its resting state).

The user-visible effect is that now, if a pipeline has not started
within 10 seconds, the 'start pipeline' component will revert back
to its regular state and we'll stop explicitly looking for the
execution.  It will appear at the next scheduled refresh of the
visible executions.