feat(exp): deployedServerGroups now grabs deployments from deploy result (#2812)
* feat(exp): deployedServerGroups now grabs deployments from deploy results

The deployment result is the place that cloud providers can put info about their deployments, but it is deeply nested on the context and hard to find. That can be eased by appending it to the info returned from the existing deployedServerGroups property/function.

* Changed generics into nested classes

I considered making this model generally useful, but after casting the context just to get a very specific path it doesn’t feel like it belongs with the more general purpose stuff under model. It actually seems to conflict with the existing stage/task/context model and would cause confusion.