feat(provider/kubernetes): Add traffic options to deploy description (#3551)
* refactor(provider/kubernetes): Improve some encapsulation

KubernetesDeployManifestDescription is a Data class, so the
member variables can be private.

KubernetesManifestTraffic is simple enough that using a builder is
less clear; give it a single-arg constructor and remove setters.

Also lightly refactor KubernetesDeployManifestOperationSpec to make
it easier to add tests in the next commit.

* feat(provider/kubernetes): Add traffic options to deploy description

In order to support better first-class deployment strategies in the
Kubernetes V2 provider, add functionality to the deploy manifest
stage.  In particular:
* Allow the services to be specified on the deploy description, rather
than only as an annotation on the manifest.
* Add a boolean to control whether to send traffic to the new workload,
defaulting to true.