chore(install): Add Arch Linux installation (#1251)
* Adapted install script for arch based distros

Not a lot of changes, since bash is bash anywhere, but this is a list:

 * Uses `who` instead of `who -m` that doesn't behave correctly on arch,
 I guess something different about how loginctl works, but couldn't
 find anything specific.
 * Uses pacman to install the needed java/ca packages instead of dpkg.
 * Removed the migration check, that is specific to past debian
 packages.
 * Uses a `sysusers` file to create system groups instead of using
 `groupadd` as suggested by the packaging guidelines.

 Tested on Arch and Manjaro, and worked fine. There may be some
 imcompatibility with other arch based distros but should cover most of
 the options around.

* chore(install): Add Arch Linux installation

Not a lot of changes, since bash is bash anywhere, but this is a list:

 * Uses who instead of who -m that doesn't behave correctly on arch,
 I guess something different about how loginctl works, but couldn't
 find anything specific.
 * Uses pacman to install the needed java/ca packages instead of dpkg.
 * Removed the migration check that was specific to past debian
 packages.
 * Uses a sysusers file to create system groups instead of using
 groupadd as suggested by the packaging guidelines.

Still mentions debian for the download location/archives since there's no
generic named file to download.

Tested on Arch and Manjaro, and worked fine. There may be some
imcompatibility with other arch based distros but should cover most of
the options around.

* Generic linux install script

This makes the old InstallHalyard less debian/ubuntu centric by adding
Arch Linux based distros to it. Moved it to a `linux` directory,
following the current standard (there's a `macos` directory)

Tested on pristine vagrant images and both install correctly. Should be
easy to adapt to other distros.

* Missed a space