feat(gcb): Add ttl to build cache entries and fall back to polling (#431)
* style(gcb): Change indent to 2 spaces

The default for igor was recently changed to 2 spaces to match our other
microservices; update the 2 gcb files that still have 4-space indents. This
commit is whitespace-only.

* refactor(gcb): Extract all interactions with GCB API to a client class

Currently GoogleCloudBuildAccount directly generates calls to the GCB
API (though uses a GoogleCloudBuildExecutor to actually execute the
requests). As we add more complex logic to GoogleCloudBuildAccount and
want to add some unit tests, it makes sense to pull the logic that is
just generating API calls into a separate class so it can be easily
stubbed.

Create a GoogleCloudBuildClient which will expose GCB API operations
and execute them for the consumer.

* refactor(gcb): Standardize return values in GoogleCloudBuildAccount

We'll standardize on having the account class accept and return GCB objects
rather than serialized objects or maps.

* feat(gcb): Add ttl to build cache entries and fall back to polling

In order to handle cases where a pubsub message is not delivered,
we'd like to fall back to polling if we haven't received an update
on a build in some amount of time.

Configure this by adding a TTL on all entries in the cache; we'll
set a shorter TTL for in-progress builds (as we expect an update)
and a longer one for completed builds (which should never have
another status change). If we check the cache and there is no build
there, fall back to an external poll, and store that result for
the same amount of time.