feat(gcb): Fetch artifacts produced by a GCB build (#2882)
* feat(gcb): Fetch artifacts produced by a GCB build

* refactor(gcb): Refactor monitor task to use generic retry logic

Now that RetryableIgorTask is generic, MonitorGoogleCloudBuildTask
can extend it instead of having its own retry logic.

* fix(gcb): Add longer timeout to GCB polling stage

We're currently using the default timeout of 1 minute, which will
likely be too short.  As a starting point, use the timeout and
backoff period we use for other CI jobs, though we can change
these in the future if needed.