feat(cf): add build info to server group (#6912)
This is similar to how it is represented in the AWS impl. We could probably standardize the attributes across providers. I’m using “ciBuild” rather than “buildinfo” because CF already has a “buildinfo” that refers to creating the droplet with a build pack. In the view I prefer to incorporate the link with the build number rather than show it explicitly. There is not a lot of horizontal real estate in the detail side panels.