fix(ci): Fix two bugs from refactor of halyard CI commands (#1300)
* fix(ci): Fix error casting CI class from daemon

It turns out that removing the objectMapper from the halyard-cli
commands broke deserializing the result from the daemon. We'll need
to continue to conver the result to the expected class.

Instead of adding back reflection, make a small enum that keeps track
of the classes relevant for each CI system, and call that enum to
do the mapping.

* fix(ci): Rename getAccounts to listAccounts

The interface method getAccounts starts with 'get' so the JSON
serializer is assuming that it is the accessor for a property;
this means that requests for jenkins accounts are returning
accounts under both 'masters' and 'accounts' because both
getMasters and getAccounts are defined.

We could add @JsonIgnore to the getAccounts interface method,
but we don't want to ignore that for gcb, and I don't believe
you can override a @JsonIgnore annotation in a subclass (or at
least this didn't work when I tried it.)

Avoid all of this by just calling the interface method listAccounts
so it's not confused as an accessor for a property, and then let
Google Cloud Build keep its getAccounts accessor.