fix(artifacts): Allow defaulting in execution artifact (#6961)
* fix(artifacts): Allow defaulting in execution artifact

Find artifact from execution doesn't use the full expected artifact
component, which means users can't assign a name or use a default
artifact. Upgrade it to the full expected artifact component so
it gets these features.

Also fix the field defaulting; currently the fields will reset if
you navigate away and back to the stage before saving.

Also add the "Artifact Status" tab to this stage so it's easier
to see the artifact that was bound.

* fix(artifacts): Hide trash icon if artifact can't be deleted

In the "Find artifact from execution" stage we need exactly one
artifact so we don't allow the existing artifact to be deleted.
The trash can is still showing up, but has no callback defined;
hide it if there is no delete callback.