chore(dependencies): upgrades to spring-boot 2 and spring-security 5 (#794)
Now uses kork BOM for dependency management.

The most significant change in here is that with Spring-Security 5 each WebSecurityConfigurerAdapter creates its own filter chain, and the first (by @Order) that matches a request will handle that request.

In kork we moved the couple of WebSecurityConfigurerAdapters to high precidence, and restricted them to only match requests to the relevant paths (/health, /metrics).

In gate we support one type of SSO / user facing auth enabled, as well as x509.

The WebSecurityConfigurerAdapter for x509 is set to match any request but at a lower-than-default precidence (2000) while the other WebSecurityConfigurerAdapters remain at order 1000 and install a request matcher that checks whether there are multiple ports configured in the web server. If so, it will only match requests where the localPort equals the server.port allowing x509 auth to handle any other API ports that are configured.