fix(*): Make config defaulting comptatible with boot2 (#69)
* fix(*): Make config defaulting comptatible with boot2

In boot2, spring.config.location overrides the default instead
of adding to it; this means that we're no longer pulling in the
default config file for the service.

We could change to using additional-location but this isn't
supported for services still on boot1. Instead, just manually
add in the default locations (which are the same for boot 1 and 2)
before /opt/spinnaker/config. Once we are fully on boot2, we can
remove the defaults and change the parameter to additional-locations.

* fix(*): Use additional-location

We can't upgrade to spinnaker-gradle-project > 6.0.0 without
being on boot2, so no need for backwards compatibility.