fix(redis): Ensure redis is started (#1309)
When installing redis on ubuntu 16.04, it is not enabled by default
(as opposed to 14.04 and 18.04 where it is). This breaks our
integration tests, and also likely confuses users who need to
manually start it even after deploying with halyard. To fix this,
ensure the service is started after installing.  (systemctl won't
exist on 14.04, so guard that statement with a || true)