feat(cats): Allow partial cache updates
Currently, cache providers (either Default or SQL) evicts all those
keys not present on the CacheResult. This prevents doing partial
updates in on-demand caching agents, which may come in handy when
the query operation is a costly one and only a subset of the
keys may actually need updating.

This patch introduces the notion of a partial CacheResult, which allows
keeping the keys not present on the CacheResult if the partialResult
flag is set to true. This change is backwards compatible, keeping
the previous behaviour (replacement strategy) by default.