fix(core): Fix halyard cli (#1330)
The boot2 upgrade removed halyard-cli as a dependency of halyard-web.
This was reasonable to do, as halyard-web does not directly depend on
any functionality of halyard-cli. Unfortunately, the halyard install
process does not separately install halyard-cli; it only installs
halyard-web into /opt/halyard but the CLI nonetheless expects to
find its jar file there.

This means that currently the halyard-cli won't run when installed
using the Halyard installer. While there's almost certainly a better
way of implementing this, just re-add the dependency for now to
unbreak halyard.