fix(kubernetes): Fix a few other bugs from client library upgrade (#3811)
V1HTTPGetAction.getPort() now returns an IntOrString. This supports
.toString() to always convert to a string, but does not support
.toInt() so we need to handle parsing as an int if desired.

The logic to serialize an IntOrString needs to know that it should only
dump either getString() or getInt() depending if it's a string or
integer, which our ObjectMapper doesn't know about. The Yaml object that
comes with the kubernetes API is configured to handle this, so just use
that instead to handle serializing.