feat(cloudformation): Make the template editor more lenient (#7290)
When paired with the changes in https://github.com/spinnaker/orca/pull/3024 and https://github.com/spinnaker/clouddriver/pull/3849, this makes it possible to use Yaml instead of Json in the editor, and it will also accept CloudFormation specific functions like `!Ref` and `!Join`. Since Json is a subset of Yaml, Json will still be accepted and parsed correctly.