feat(bakeManifest): add kustomize support (#7342)
adds support for kustomize in the UI. the main user input is a single
kustomize file which tells the bakery what it needs to do to render the
manifests. broke the helm config form into a new component as well.
putting this behind a feature flag because the backend isn't complete
yet. will remove once we merge the backend.