fix(buildtool): log stderr of `gcloud compute instances describe`
By default, it was getting joined with stdout and then we tried to parse
it as JSON, which didn't work since it isn't, in fact, JSON.

My test run didn't actually have any of this stderr output, so I can't
say for certain that it actually works. But it doesn't make the
no-stderr case fail, so we're at least no worse off.