fix(newrelic): Make New Relic canary graphs work (#469)
This is a small patch to make sure the graph view for New Relic canaries
doesn't crash.  This will soon be extended to properly handle UI
configuration of New Relic canary configs, but wanted to get this in
first, since crashes are bad.

Ref:  https://github.com/spinnaker/spinnaker/issues/4389

Kudos to @imakewebthings for pointing out the underlying cause.