fix(rollback): capture the source capacity before anything else
Does not entirely address this issue but makes it less likely to happen:
https://github.com/spinnaker/spinnaker/issues/4895

The "danger zone" is when a resize of the source server group happens
between the start of the rollback and the moment we take the capacity
snapshot. This change should reduce that "danger zone" from minutes to
seconds.