fix(core/pipeline): Fix execution details chevron for grouped stages (#7421)
The Execution Details link in the pipeline view doesn't work if the first stage in the pipeline is a group. I think this was by design, but it has been a major cause of confusion for our users, so I think it makes sense to just expand the first stage with the first substage selected.