fix(kubernetes): Add networking.k8s.io/v1beta1 ingress versions (#4056)
* refactor(kubernetes): Pull api versions to a set

A number of different places check the API version of a manifest
against a set of supported ones using a log if condition. Replace
these with a statically-defined set of supported versions and a
check of whether the manifest's API version is in the set.

This commit doesn't change any of the supported versions.

* fix(kubernetes): Add networking.k8s.io/v1beta1 ingress versions

In Kubernetes 1.20, Ingress will stop being served from the
extensions/v1beta1 API group and will only be served from the
networking.k8s.io/v1beta1 API group (which has been available
since 1.14).