refactor(build): Update determine push/pull URL to take the origin
We're passing a full repository to determine_pull_url and
determine_push_url, but these functions only need to know the origin.
To allow these to be re-used in cases where we only have an origin,
change the functions to accept an origin and expect callers to pass
one in.