fix(build): Fix repository comparison to handle ssh/https difference
When checking to see if a directory contains the expected git
repository, we compare the origin against the expected origin. This
often doesn't work as we have logic to auto-translate an origin from
https to SSH before cloning, so existing repositories will often have
ssh origins even if the configured origin is https. Fix this by
considering repositories equivalent if the origin is either the
expected origin or the result of determine_pull_url on the
expected origin.