fix(build): Fix repository comparison to handle ssh/https difference (#4975) (#4976)
* refactor(build): Update determine push/pull URL to take the origin

We're passing a full repository to determine_pull_url and
determine_push_url, but these functions only need to know the origin.
To allow these to be re-used in cases where we only have an origin,
change the functions to accept an origin and expect callers to pass
one in.

* fix(build): Fix repository comparison to handle ssh/https difference

When checking to see if a directory contains the expected git
repository, we compare the origin against the expected origin. This
often doesn't work as we have logic to auto-translate an origin from
https to SSH before cloning, so existing repositories will often have
ssh origins even if the configured origin is https. Fix this by
considering repositories equivalent if the origin is either the
expected origin or the result of determine_pull_url on the
expected origin.