feat(provider/aws): Functions (listing and searching) (#7568)
* .

* feat(provider/aws): Function listing and searching functionality

This change adds a new tab for functions and facilitates listing existing functions and searching functions listed.
Corresponding provider specific changes are included in this change for AWS.

Authors: @sidmuls and @nabebe

* feat(provider/aws): Function listing and searching functionality

This change adds a new tab for functions and facilitates listing existing functions and searching functions listed.
Corresponding provider specific changes are included in this change for AWS.

Authors: @sidmuls and @nabebe

* refactor(provider/aws): Converting Function class component into function component.

This is a refactoring change as requested by reviewer @caseyhebebrand

Author: @sidmuls

* refactor(provider/aws): Refactoring groupings into a separate component.

This change is for refactoring requested by reviewer @caseyhebebrand.

Author: @sidmuls

* refactor(provider/aws): Removing unnecessary state from FunctionGroupings.

This change is a refactoring requested by @caseyhebebrand

Author: @sidmuls

*  refactor(provider/aws): removing extra line. Tthis is actually for retrying tests on TravisCI

* refactor(provider/aws): Adding tests for FunctionFilterService and changes as requested by @christopherthielen

Author: @sidmuls

* refactor(provider/aws): removing unnecessary field from IFunction

* refactor(provider/aws): adding flag to turn off functions feature