feat(cfn): request to delete CFN changeset if empty (#3234)
* feat(cfn): request to delete CFN changeset if empty

This patch requests the deletion of the CFN change set if it's empty.
The rationale behind it is that whenever a change set does not contain
changes to the CFN template, the AWS API marks the changeset as FAILED.
However, in a CD world, an empty change should not be considered a
failure, and instead continue the normal execution flow.

In this patch we detect that the change set is empty by the status
message and request it's deletion accordingly. This is implemented as
a separate task so it can be reused in different scenarios planned for
the future (e.g. configuration option to delete all changesets that
have been executed).

* Address comments