fix(notifications): Evaluate SpEL expressions in pipeline notifications (#3280)
If a pipeline notication uses a SpEL expression which can't be evaluated at pipeline submission time (e.g. `${authentication.user}`)
that expression will not be evaluated prior to sending a notication. And the notification will have the raw expression present.

Ironically, app-level notifications don't suffer from this illness - they are properly evaluated and resolved prior to sending the notication.
This change adds explicit eval of notifications on the current execution