fix(docker): Fix sorting array of non-comparable types (#4183)
* test(docker): Add tests on docker registry tag fetching

There are currently no tests that the fetching of docker registry
tags and the sorting of these tags by creation date works properly;
add some tests before touching that code.

* fix(docker): Fix sorting array of non-comparable types

getCreationDate returns an Instant; when the call fails, we're
returning a default value as a Date. This means we can end up with
an array with a mix of Instant and Date objects, which don't
together obey the general Comparable contract.

Fix this by returning an Instant from the catch block; also pull
the error handling logic into a CompileStatic function so we
get some additional type safety.

* refactor(docker): Pull key-based sorter into its own class

When sorting docker tags by creation date, we are effectively doing
a Schwartzian transform to map each item to a key, sort by that key,
then re-map back to the original item. Rather than have this ad-hoc
in Groovy, make a simple Java class that handles this algorithm and
that can be tested. This also allows us to mark the method that
contained this logic as CompileStatic, as we've removed the non-safe
operations into Java (and made them safe).

Right now, leaving the class alongside the single place it is used,
but if we need this in other places we can move it to a more general
location.