refactor(*): Move artifactory integration into module (#562)
This PR starts the effort of making Igor build system integrations built as
plugins. The first step is to break the integrations into modules so we have
a better sense of what the right abstraction are.

I bit this effort off purely to get more experience with real-world plugin
implementations that aren't stages. Figured Igor needed some organizational
love anyway, and since it's such a small service, it could be completed
reasonably quickly (if no plugin bugs arise :sweat_smile:)