fix(tests): Use webpack-dev-middleware for functional tests (#7687)
* fix(tests): Use webpack-dev-middleware for functional tests

The new CSS loading plugin TypedCssModulesPlugin is causing
webpack-dev-server to compile the application twice; the first
generates the co-located .css.ts.d files and the second recompiles
because of a change in sources.

This is breaking the functional tests, as the server reports healthy,
causing the tests to start, but then immediately recompiles. This means
that the first tests needs to wait for this second compilation pass,
which is leading to intermittent (but frequent) errors.

The fix here is to turn off hot reloading of the app. There does not
appear to be a way to do this using webpack-dev-server, so switch to
using webpack-dev-middleware to set up the test server.

* style(tests): Fix a few issues with prior commit

* Remove extraneous console.log
* Restore accidentally removed repoRoot from constructor and pass
it along to the webpack config
* waitOn returns a promise itself so there's no need to explicitly
create one

* style(tests): Use require for consistency