feat(queue): use keiko 3.1.0 with eventless metrics (#3349)
* feat(queue): use keiko 3.1.0 with eventless metrics

The queue event listener and queue state poller are now native in keiko, and the events no longer go through Spring.

AtlasQueueMonitor used to do listen to queue events, poll the queue state and check for zombies. It is now only responsible for the zombie check, so let's rename it accordingly to ZombiePipelineCheckingAgent.

Since it looks more like an actual agent now, I think it makes sense to make it extend AbstractPollingNotificationAgent like the other agents, so it doesn't need to do its own scheduling and locking anymore. Which is why we can remove that test case from ZombiePipelineCheckingAgentTest

Also:
- rename ZombiePipeline... to ZombieExecution...
- pull keiko version post boot 2.2 revert