refactor(artifacts): Make artifact constructors private (#463)
The all-arg and no-arg constructors for Artifact and ExpectedArtifact
are marked as deprecated, as consumers should be constructing artifacts
by using the lombok-generated builder class.

Remove the no-arg constructors entirely, and make the all-arg
constructors private.  The all-arg constructors are used by the
inner builder class; had we omitted the explicit AllArgsConstructor
annotation, the Builder annotation would have generated a
package-private all-args constructor for us. The goal of explicitly
adding the annotation is so that we can make the constructor private.