refactor(artifacts): Rename ArtifactResolver to ArtifactUtils (#3375)
* refactor(artifacts): Rename ArtifactResolver to ArtifactUtils

This class does a lot more than resolve artifacts, and an upcoming
commit will pull out the resolving-specific functionality into its
own, smaller class.  In order to leave the ArtifactResolver class
name available for that new class, rename this to ArtifactUtils.

In general, *Utils classes tend to get a lot of functionality that
should live in better purpose-built classes so are often not a great
idea. In this case, the class already has a bunch of functionality so
the name accurately describes it; the upcoming commit will pull out
one part of the functionality and hopefully future refactors will
further pare down this class.

* refactor(artifacts): Convert find artifacts test to java

An upcoming commit will require some changes to this test; as it's
easy to convert it to java (and makes the changes easier) just convert
it to java here.