fix(test): Fix flaky queue integration test (#3382)
The test are starting to fail quite frequently with an assertion
failure "Pipeline did not complete".  This is because we assert
that an execution has started within 5 seconds of requesting it,
which appears to fail frequently when running in constrained CPU
environments.

Double the timeout to 10 seconds to alleviate the issue, though
a better fix would probably be to avoid having an arbitrary timeout
at all.  (Perhaps the test function could wait until the queue
is quiescent before proceeding?)