refactor(kubernetes): Move a few classes to better packages (#4284)
* refactor(kubernetes): Move a few classes to better packages

KubernetesV2SearchProvider should be in the same package as all
the other Kubernetes providers; move it there.

KubernetesAccountResolver is really just a helper class for the
classes in .*kubernetes.v2.caching.view.provider; move it there
so it can be made package-private and only visible to those
classes.

GlobalKubernetesKindRegistry is closely associated with the
account credentials and should live in the same package (as
the account-specific kind registry does).

* refactor(kubernetes): Remove CacheUtils from handlers

The hydrateSearchResult function in all the kubernetes handlers
accepts a KubernetesCacheUtils, but never uses it. Remove the
unneeded parameter from these functions.

* refactor(kubernetes): Make classes package-private

The KubernetesAccountResolver and KubernetesCacheUtils classes
are now only used within their package (after the prior commits);
make them package-private.

* refactor(kubernetes): Undo package-private of CacheUtils

Let's first try to understand better exactly how people are using this
in layered builds.  This is one of the unfortunate issues with having
poorly-defined extension points and every class public---anyone can
use effectively any class in layered builds.