refactor(kubernetes): Clean up some account resolving code (#4285)
* refactor(kubernetes): Remove isAccountRelevant and makesLiveCalls

Both implementations of KubernetesV2AbstractManifestProvider separately
call isAccountRelevant and makesLiveCalls to determine if an account
exists. This duplicates work and also adds an unnecessary layer of
indirection; instead juts have them get the optional account and
ask the account (if present) directly what they need to know.

This commit makes resourcePropertyResolver protected in the parent
class so that subclasses can use it, but a later commit will just
push it directly down to the subclasses.

* refactor(kubernetes): Remove getKindRegistry

Just as with the prior commit, we don't need the account resolver
anymore by the time we're calling this as we already have valid
credentials; just call credentials.getKindRegistry() to get the
kind regsitry.

* refactor(kubernetes): Pass credentials instead of account name

In a number of the functions in the KubernetesManifestProvider (and
LiveManifestProvider) we pass around an account name even when we
already have the credentials; to avoid having the function we call
need to resolve the credentials again, just pass the credentials.

To account (no pun intended) for cases where we need the name of the
account (generally for logging), add a Getter to get the account name
from the credentials.

This allows us to remove getRegistry from the base manifest provider
as it is no longer used.

* refactor(kubernetes): Add NonnullByDefault to account resolver

Also remove an unused Slf4j annotation

* refactor(kubernetes): Push account resolver to implementing classes

The KubernetesAccountResolver is never actually used in the abstract
base class; move it to the implementations. Also rename it in the
instances from resourcePropertyResolver to accountResolver as that is
a better name.

Finally, mark buildManifest as static as it does not depend on any
instance properties.

* refactor(kubernetes): Eliminate abstract base class

At this point, all the KubernetesV2AbstractManifestProvider is
doing is supplying a single static method. Rather than using
inheritance to share it between the two implementations, we'll
use composition.

* Change the two concrete manifest providers to directly
implement ManifestProvider<KubernetesV2Manifest>
* Rename KubernetesV2AbstractManifestProvider to
KubernetesV2ManifestBuilder as it helps build a manifest
* Call the static method in KubernetesV2ManifestBuilder from
the manifest providers

The abstract class being eliminatd is not one that should be
used by people extending Spinnaker; the only reason there is a
base class at all was to sit on top of live- and non-live- manifest
providers.

There may be a better place for the single method in
KubernetesV2ManifestBuilder but for now I'll leave it there and
will leave finding a better place to a future refactor.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>