feat(clouddriver): make selector accept configurable criteria (#1032)
Historically, ClouddriverServiceSelector was hardcoded to look only at the user origin (coming from the X-RateLimit-App header), and was able to look up a different url for clouddriver based on the value of that header.

With this PR, ClouddriverServiceSelector accepts a configurable SelectableService for which we can pass any number of criteria, similar to the existing pattern for e.g. orca and rosco. To avoid a breaking change, GateConfig will try to detect the previous clouddriver selector style (a `dynamicEndpoints` map) and configure an equivalent SelectableService. Additionally, we support a dynamic config to turn off routing at runtime if needed.

Note that RequestContext does not currently contain most criteria that gate can be configured to route on, will change this in an upcoming PR.