refactor(kubernetes): Simplify status handlers (#4303)
* test(kubernetes): Add test for progress deadline exceeded

The test for progress deadline exceeded only considers the case
where the deployment reports it is available; also add a case where
the deadline is exceeded and the deployment reports it is not
available.

* refactor(kubernetes): Improve Status and Condition classes

There are way too many ways to create and mutate these classes,
as both have fully public fields as well as setters; in the case
of Status there are also custom setters, and in the case of Condition
there is additionally a builder.

Clean this up as follows:
* Update the custom setters (ex: failed, unstable, etc.) in Status
to create a new condition and assign it to the appropriate value
instead of mutating the existing condition. (This fixes a null-safety
issue as some of these fields can be null, if unknown() has been
called.)
* This now allows us to make Condition immutable, as we never change an
existing condition. Make it immutable by removing the setters and
marking the fields final. Replace builders with a constructor (which is
reasonable given that we have only two fields with different types). As
it's common to create a condition with a null message (because it shows
up in defaults for Status) make a convenience factory method withState
that returns a Condition with null message.  (Because the class is
immutable we can just return a pre-created static Condition in this
case.)
* Add a defaultStatus() factory method to Status; it's the same as the
constructor new Status() but I think makes the intent much more clear
when reading the Handlers; that is, start with some default status and
mutate it if the desired status differs from the default.  I didn't make
the no-arg constructor private because I didn't want to needlessly break
people who used it in their own custom handlers, but ideally they'd
migrate to defaultStatus() for clarity.
* Add a stable() mutator to replace the one case where we explicitly
called setStable()
* Add nullity annotations.  Many things are nullable; the new
annotations just document that existing behavior.

* refactor(kubernetes): Simplify status reporting in handlers

The logic to report workload status is fairly complex to follow;
one of the reasons it's complex is that in general we create an
empty Status object (which defaults some fields) at the beginning
of the status function and mutate it throughout the functions,
sometimes returning immediately after the mutation but sometimes not.

In order to understand what we're returning at any given point, we
need to read the entire function up to that point to see if any
of the prior conditions wrote to the result variable.

Where possible, simplify this by just creating a new Status inline
at the point of return (and change from using new Status() to the
factory method Status.defaultStatus()). This is possible in any
cases where there are no mutations of result earlier in the function
that don't immediately return. By doing this analysis now once, we
make it easier for future readers of the function to see what is
returned in each case.

This commit only touches the handlers where it was trivial to make
this change (ie, we never mutate the result without immediately
returning it).

* refactor(kubernetes): Simplify Deployment and ReplicaSet handlers

Now that there is a robust test suite on the Deployment and
ReplicaSet handlers, refactor these to make the control flow
more clear when determining status. This commit only makes
changes that do not affect results (as seen by the lack of
changes to the tests in this commit).

* fix(kubernetes): Fix ready condition check for replicaset

We currently check the available replicas count before we check
the ready replica count; this isn't correct as there will always
be more ready replicas than available replicas. (The definition
of an available replica is one that has been ready for at least
minReadySeconds.) Switch the order of the check so that we
generate the correct message in each case.

* fix(kubernetes): Simplify status messages

When a replicaset is unstable and unavailable we use slightly
different wording in the unstable and unavailable message even
though the cause is the same. This isn't complicates both the
code and the user's parsing of messages (as they need to parse
both to realize they are the same error).

Simplify this; the tests show exactly where the text of errors has
changed and shows that the actual stable/unstable status does not
change.

Also, push the Deployment's UnstableReason up to the top level so we
can re-use it for replica sets.

* fix(kubernetes): Don't return early after setting failed flag

When a deployment reports that it is failed we exit early instead
of checking the remaining condition. This means that we don't always
properly set the other flags of the status in this case; update the
logic to set the other flags.

The logic in orca to decide how to handle the status of a manifest
checks both the stable and the failed flags (so we won't risk breaking
it by sometimes setting a deployment to unstable when it is also
failed, which we already do in some cases anyway if the condition
reports unavailable).

* fix(kubernetes): Fix copyright header

I just realized I checked in an incorrect copyright header.

* refactor(kubernetes): Make condition constructors private

As I was thinking about how these changes might affect users with
custom handlers, I realized that there's actually nothing anyone
can do with a Condition directly anymore.  Since there are no
setters on Status, you can't pass a condition to a status; it's
up to Status to create this based on your message.

Based on this, make the Condition constructor/factory private to
simplify the API. This doesn't break anything additional as there's
nothing you could do with this condition.

Also making the class final (as making the constructor private
effectively does this).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>