fix(kubernetes): Fix and simplify manifest stability logic (#3413)
* test(kubernetes): Add tests to WaitForManifestStable and Manifest

There are some existing tests in groovy on WaitForManifestStableTask
but they are missing some edge cases that will prove important in the
next commit.  Add a new suite of tests in Java; these duplicate the
Groovy tests and also add more test cases. We'll wait to delete the
Groovy tests until after the refactor in an upcoming commit just to
get a bit more validation that it doesn't break anything.

Also to prepare for some upcoming changes, add tests to the Manifest
class to ensure we can correctly deserialize manifests.

* fix(kubernetes): Fix and simplify manifest stability logic

The logic in WaitForManifestStable is quite complex; in particular
it separately distinguishes true, false, and null on the manifest's
status field and has different behavior in each case. Likely due to
this complexity, there are a number of bugs that were surface when
writing tests in the last commit.

Fix these bugs, which also allows us to stop treating a null status
as different from false; the cases where we did were just what was
leading to the bugs noted earlier.

The tests should clearly show the changes, but in particular:
* The status is fixed in a couple of cases where we either failed
too early when one manifest was failed, or eventually returned success
incorrectly even though a manifest had failed on an earlier poll.
* Duplicate log messages are removed; if a manifest has failed, we
now just log that it failed instead of also logging that we are
waiting for it to stabilize.

* refactor(kubernetes): Delete unused ObjectMapper and groovy test

The ObjectMapper is unused. The groovy test was superseded by the
Java tests added a few commits ago, but was kept around temporarily
to give additional validation to the refactor in the last commit.

* refactor(kubernetes): Make Manifest classes final and immutable

Let's make the Manifest class and contained classes immutable.

The changes to WaitForManifestStableTaskTest are just to take advantage
of the more convenient constructors/builders that were created on
those classes; none of the actual test results change.

* refactor(kubernetes): Default null entries in manifest

Add nullity annotations to the Manifest class (and contained
classes) and default any null entries on deserialization.

The fact that none of the tests on WaitForManifestStableTask
broke means that switching from having a null status to a false
status has not changed any observable behavior.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>