refactor(swagger): Remove groovy dependency from kork-swagger (#515)
After submitting the last PR, I realized that there is a way to
exclude groovy from Swagger annotations without forcing it as
a dependency.

In general I'm not a fan of reflection as it makes it really hard
to perform static analysis and is hard to maintain. But this may
be the one time in my life where it is actually the right solution
to the problem.