fix(peering): Don't use `null` partition on local db (#3466)
For historical reasons, `partition=null` is set on all current executions (since sqlrepo runs without a partition).
So when checking if an execution is local we need to check if `parition == MY_PARTITION` OR `partition is null`.
However, this check is not correct when copying from the peer. From the peer we need to get `partition == peerId or partition is null`,
but locally it must always be `partition == peerId` because we force set the partition during copy.
Without that we would think that we need to delete all our local executions since they would not show up in the peer query
This change fixes this issue