config(core): Push halyard config defaults down to base (#3482)
* config(core): Push halyard config defaults down to base

This change pushes a number of config settings that were in the
Halyard-specific config file down to the base config so they apply
to all uesrs.

The one potentially breaking change is that echo used to default to
disabled but now defaults to enabled. Given that most users use
Halyard and netflix uses echo, I suspect that nobody was relying on
this to default to off. We can document this as a change in the
release notes (as it's easy to override if desired).

* config(core): Update bakery defaults

Push the defaults for extractBuildDetails, roscoApisEnabled,
and allowMissingPackageInstallation down to the properties class.
(As defaults were there anyway, there is no need to specify them
in clouddriver.yml.)

This makes the following changes:
* Removes references the services.rosco.* settings for two flags;
only the Halyard config referenced them, but Halyard never wrote
out those values.
* Sets extractBuildDetails to default to true; this has always been
the case for Halyard users.

* config(core): Update redis config setting

This change updates the core redis config setting to read from
services.redis.baseUrl then default to localhost:6379.

This does remove looking at the REDIS_URL environment variable. I
strongly suspect this isn't being used by anyone (as definitely
it is not set by Halyard users), though would like to confirm with
Netflix before merging.

* feat(config): Add readme to halconfig directory

* config(core): Add readme to halconfig

Also remove the redis config from the halconfig, which I forgot to
do in the last commit when adding it to the base config.

* config(core): Fix README