fix(metrics): record drift from last resource check (#976)
* fix(metrics): record drift from last resource check

Calculate the resource check drift metric (keel.resource.check.drift)
by using the last time resource check succeeded as the start time.

Previously, it was calculating based on the beginning of the actuation
loop, which would drop to zero when the loop restarted due to timeout.

Fixes #950

* fix(pr): move ResourceCheckCompleted, remove ScheduledResourceCheckStarting

Remove ScheduledResourceCheckStarting since it's no longer used.

Move ResourceCheckCompleted to a separate file.

* fix(pr): add missing import