refactor(kubernetes): Clean up runAndRecordMetrics (#4495)
* test(kubernetes): Add tests to metric recording on kubectl calls

The runAndRecordMetrics funcgtion in KubernetesV2Credentials is
very complex and has no tests. Before simplifying it, add some
tests to document and validate the current behavior.

This commit also updates the constructor of KubernetesManifest to
set the metadata to an empty map. In real code, we always create a
KubernetesManifest by deserializing, so the metadata is set to a
map of whatever was on the incoming manifest. But it's often useful
in tests to be able to create a manifest and set a few fields on it;
before this change, this wasn't possible because metadata was always
null. There were functions to set various keys on the metadata, but
these would all fail due to the null metadata.

* refactor(kubernetes): Clean up runAndRecordMetrics

The control flow in runAndRecordMetric is really confusing, in
particular in that we re-throw and return from a finally block.
Clean this up so that it uses better style.

* fix(kubernetes): Fix some issues with kubectl metrics

Cleaning up the code in the last commit surfaced a few issues
that this commit cleans up.

(1) We should add a failure tag if a KubectlException is thrown
(2) Metric tags are intended to be short descriptors, so we shouldn't
add the text of an error message. Remove that message from the cause
tag and just include the class of the exception.
(3) There's no reason to separately tag 'kind' and 'kinds' depending if
there are more than 1 or not; just always add 'kinds' with the full list
of kinds. Also sort it for consistency in the metric tags.

There's probably a bit we can do to futher clean up the exception handling
(deciding when to wrap the exception, etc.) and deduplicate the two catch
blocks but leaving that s is for now.

* refactor(kubernetes): Don't needlessly qualify KubectlException

Co-Authored-By: Maggie Neterval <mneterval@google.com>

* refactor(kubernetes): Fix comment and assert on class

One of my comments was nonsensical; fix it, and also assert on the
exception class in that test.

Co-authored-by: Maggie Neterval <mneterval@google.com>