Schema upgrade (#974)
* feat(migration): Basic implementation migrating spec schemas

* feat(migration): Typed version of SpecMigrator

Suggested by @lpollo. I think I like this a lot better. I'd like to take it further by establishing a direct relationship between `ResourceKind` and the `ResourceSpec` subclass.

* chore(tests): Sanity checked test usage of ResourceTypeIdentifier

There were a couple tests that were creating a `ResourceTypeIdentifier` and wiring it in, but then it never got used, being ignored in favor of `DummyResourceTypeIdentifier`.

* refactor(kinds): Constants for each supported kind