refactor(kubernetes): Improve error abstraction in kubectl (#4498)
* refactor(kubernetes): Improve error abstraction in kubectl

When parsing the output of kubectl as JSON, a few different kinds of
errors can happen:
(1) An IOException if we encounter an error reading from the output
stream
(2) A JSONSyntaxException if the call to gson.fromJson fails
(3) An IllegalStateException if the JSON format is unexpected (ex:
if we get a list back instead of an object and call .beginObject()
on it).

Currently we're propagating all these exceptions as is. Let's continue
letting the IOException propagate, as this can be handled by the job
executor in the same way it handles any other error reading output.

Let's wrap the JSONSyntaxException and IllegalStateException in a
KubectlException to provide relevant context. This is how we handle
a JSONSyntaxException other places in this file when we parse
non-streaming responses. (Potentially at some point this could be
a specific subclass of KubectlException but for now just being
consistent with what happens elsewhere in the file.)

This means that now the error surface out of the KubectlJobExecutor
for a JSON parse exception will be the same regardless of whether
this happened in the streaming or non-streaming parser.

* refactor(jobs): Throw exceptions appropriate to the abstraction

The JobExecutor class throws RuntimeException in a few places; let's
replace this with throwing a JobExecutionException with information about
the job that failed.

Let's also take the opportunity to document the contract about when the
JobExecutor throws an exception vs. when it returns a JobResult with a
status of FAILED.

* refactor(kubernetes): Remove NoResourceTypeException

This exception was added in PR2344, with the goal of in some cases
catching exceptions where a type was not recognized by the API
(for example if a built-in kind is not present on an old Kubernetes
cluster). It appears, however, that the change in that PR never
actually worked, as all relevant calls to kubectl go through
runAndRecordMetrics, which wraps any exception that is not a
KubectlException in a KubectlException. So the catch blocks
added in that PR will never be triggered as we'll always see
a KubectlException at that point.

Probably what should have been done is to make NoResourceTypeException
extend KubectlException (or potentially to have it special-cased in
runAndRecordMetrics). In the intervening time since that PR, we've
added a check the first time we try to read a kind to make sure that
it's readable (mainly for permissions issues, but it would also
address this case), so I'm not sure there's much value in actually
fixing this. (In any case, the behavior hasn't worked in the two years
since it was added and we have not heard reports of people running into
unrecognized kinds, except briefly when the readable check noted earlier
was broken.)

This commit then does the following:
(1) Removes all the catch blocks for NoResourceTypeException that are
unreachable
(2) Removes NoResourceTypeException itself; given that calling code is
not distinguishing this particular exception, there's not much benefit
to throwing it vs. a generic KubectlException (particularly given that
we're relying on string parsing to do so, and that we're only throwing
it in some of the methods, not others).
(3) It makes the constructor of KubectlException package-private instead
of protected.

* refactor(kubernetes): Fix a few places where we catch KubectlException

In the metrics caching agent, we catch a KubectlException and dig into
the message to see if it's because metrics aren't available, then return
an empty collection in that case. Let's just have topPod directly handle
that so we don't leak the implementation detail of using kubectl (and
the specific message); this makes sense for the other two places we call
this function as well.

In KubernetesV2CachingAgent, we catch, log, then re-throw an exception.
This is a vestige of when we used to just return an empty list after
logging, but that was change to propagate the exception, so there's
not really much additional benefit from also logging.

* refactor(kubernetes): Don't wrap kubectl exceptions

Let's no longer wrap all exceptions that pass through runAndRecordMetrics
in a KubectlException. While it does have the benefit of providing
some context at that level of abstraction, it is confusing that we
sometimes directly throw that in the KubectlJobExecutor and sometimes
create one by wrapping it here.

I think at some point we may want to create a new KubernetesAPIException
here that wraps anything below (either a KubectlException of a client
library exception), but for now let's just propagate whatever was
thrown below.

The one thing to be careful about is that this change doesn't break
places where we're catching a KubectlException. I've looked through
and in all cases the code is catching to deal with some expected
(but non-successful) kubectl call (ex: a missing resource, logs
not there, etc.). All of these remain KubectlExceptions (and in
some cases we should potentially even make more specific ones and
catch that specific one). What we'll no longer be catching is things
like a JobExecutionException where we fail to run the job, etc, which
I think actually is better in this case.

* test(kubernetes): Fix tests from last commit

I broke the tests I added yesterday, which I guess means they work.
Update them now that we aren't wrapping all exceptions.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>