refactor(kubernetes): Clean up KubernetesArtifactConverter (#4502)
* refactor(kubernetes): Make KubernetesCoordinates immutable

Luckily there are no existing uses that mutate it, so we can greatly
reduce its API by making it a Value class.

* refactor(kubernetes): Clean up KubernetesArtifactConverter

There are a bunch of functions there that are no longer used, so
we can significantly clean up that class (and implementations).
Also move getAccount to the only place where it is used.

* refactor(kubernetes): Make implementations of converter singletons

It's confusing how we store an instance of each artifact converter on
each KubernetesResourceProperties instance. The properties should only
care whether the resource is versioned; the deployer logic can figure
out what artifact converter to use based on whether the resource is
versioned (and whether it has been overriden by an annotation).

The goal here is less to reduce the number of created objects (which
is a side benefit, though probably minimal), but more to keep
KubernetesResourceProperties more focused, with the goal of one day
maybe merging it with KubernetesKindProperties so we don't have so
many of these classes describing kinds floating around.