chore(GHA): sleep for 5 minutes before running bumpdeps (#601)
* chore(GHA): fix the indentation

This is the way most of the rest of the files do it. There was a variety of indentation levels in this file (2 spaces, 3 spaces, 1 space), so use 2 spaces everywhere and don't indent list bullets.

* chore(GHA): sleep for 5 minutes before running bumpdeps

On the last kork bump, 4 of the 14 builds failed because the new artifact wasn't available yet. I don't think speed for the bumps is generally that critical, so let's just play it safe and use five minutes?

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>