chore(dependencies): Unpin org.jetbrains.kotlin (#772)
Bumping kork on front50 is failing because at some point recently
we upgraded kotlin, and it looks like it has a new artifact that
isn't in the pinned version.

I'm not sure why the pin was added initially, but the tests all
pass when it is removed. When this is removed, kork can successfully
be bumped.