fix(plugins): allow plugin dependencies (#123)
Now that the dependencies are put in 'lib' and used we can no longer load plugins. This is because the provided (compileOnly) dependencies like kork and spin services are packaged into the plugin and there are classpath issues (duplicate ExtensionConfiguration for example). This change bundles only implemenatation (compile is included) and runtime dependencies. Strangely the "canBeResolved" flag needs to be set on these configs, and a copy of them has to be made first or you get errors about it when running tests.
I also added the dirs containing the jars to the 'libsDirs' section of the .plugin-ref so that the dependencies are used in debugging mode too.