fix(travis): Limit the number of returned builds (#670)
* fix(travis): Limit the number of returned builds

If you have builds that creates very large logs, the number of builds returned can cause issues or timeouts in the Travis API. Limiting the number of builds seems to mitigate the issue.

Drawbacks: This will limit the number of builds you'll see in the build drop down of a manual execution to 10 (but it won't affect the build monitor at all as it uses another endpoint).

* Make the build limit configurable per Travis host

Co-authored-by: Gal Yardeni <55253849+gal-yardeni@users.noreply.github.com>