fix(titus): custom job timeout should match task timeout (#3624)
There is a lot of confusion for users when they specify a timeout on the batch job on Titus if that timeout is larger than 2hrs.
In those cases, if the user forgets to update the `stageTimeoutMs`.
This change allows the job runner to "hint" how long the task will run, if the hint is present the `WaitOnJobCompletionTask` will wait that amount of time + small buffer instead of the default 2hrs