chore(azure): Default storageContainerName (#789)
This defaulting was happening in Halyard, and will be removed from
kleat (which does not default parameters that are not supplied and
instead leaves this defaulting to the consuming service).

In order to support kleat, add the defaulting here; this won't affect
existing Halyard users as Halyard will continue to write this field
(meaning the deafult won't be used for them).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>