fix(health): OutOfService is a Down health state (#3638)
* refactor(health): use clouddriver's HealthState model

Behavior should be unchanged

* fix(health): OutOfService is a Down health state

When we're waiting for up instances, OutOfService was considered neither Up nor Down, which means that as long as some other health was Up, having one OutOfService would not keep the task waiting.

With this change, we consider OutOfService a Down status, which means that someAreUpAndNoneAreDown will return false, which will keep the task waiting.

* fix(test): fix WaitForUpInstanceHealthTaskSpec

* chore(javadoc): turn comment into javadoc in HealthState