fix(bake): Bake stage output may contain output from unrelated stages (#3652)
As @luispollo notes in the code I removed, the current implementation may include output from completely unrelated bakes that are running at the same time as the bake stage in question. I fail to see any reason this can be correct.

After applying this fix, each bake stage will contain the bakes created in that bake stage, and nothing else.

Coupled with [this PR](https://github.com/spinnaker/orca/pull/3651), this fix will make the output of pipelines with multiple bake stages correct, as all the `deploymentDetails` will be merged instead of overwritten, and thus contain all images created, and not just an arbitrary list which is timing dependent.

Co-authored-by: Mark Vulfson <mvulfson@netflix.com>